USBDevice.h: cleanup

+++ USBDevice.h
@@ -339,21 +339,21 @@ public:
     
     template <typename... T_Args>
     auto read(uint8_t epAddr, T_Args&&... args) {
-        const _EndpointInfo& epInfo = _epInfo((uint8_t)epAddr);
+        const _EndpointInfo& epInfo = _epInfo(epAddr);
         _Interface& iface = _interfaces.at(epInfo.ifaceIdx);
         return iface.read(epInfo.pipeRef, std::forward<T_Args>(args)...);
     }
     
     template <typename... T_Args>
     void write(uint8_t epAddr, T_Args&&... args) {
-        const _EndpointInfo& epInfo = _epInfo((uint8_t)epAddr);
+        const _EndpointInfo& epInfo = _epInfo(epAddr);
         _Interface& iface = _interfaces.at(epInfo.ifaceIdx);
         iface.write(epInfo.pipeRef, std::forward<T_Args>(args)...);
     }
     
     template <typename... T_Args>
     void reset(uint8_t epAddr, T_Args&&... args) {
-        const _EndpointInfo& epInfo = _epInfo((uint8_t)epAddr);
+        const _EndpointInfo& epInfo = _epInfo(epAddr);
         _Interface& iface = _interfaces.at(epInfo.ifaceIdx);
         iface.reset(epInfo.pipeRef, std::forward<T_Args>(args)...);
     }
@@ -524,15 +524,15 @@ private:
     
     template <typename T_Dst>
     void read(uint8_t epAddr, T_Dst& dst, Milliseconds timeout=Forever) {
-        const size_t len = read((uint8_t)epAddr, (void*)&dst, sizeof(dst), timeout);
+        const size_t len = read(epAddr, (void*)&dst, sizeof(dst), timeout);
         if (len != sizeof(dst)) throw RuntimeError("read() didn't read enough data (needed %ju bytes, got %ju bytes)",
             (uintmax_t)sizeof(dst), (uintmax_t)len);
     }
     
     size_t read(uint8_t epAddr, void* buf, size_t len, Milliseconds timeout=Forever) {
-        _claimInterfaceForEndpointAddr((uint8_t)epAddr);
+        _claimInterfaceForEndpointAddr(epAddr);
         int xferLen = 0;
-        int ir = libusb_bulk_transfer(_handle, (uint8_t)epAddr, (uint8_t*)buf, (int)len, &xferLen,
+        int ir = libusb_bulk_transfer(_handle, epAddr, (uint8_t*)buf, (int)len, &xferLen,
             _LibUSBTimeoutFromMs(timeout));
         _CheckErr(ir, "libusb_bulk_transfer failed");
         return xferLen;
@@ -540,14 +540,14 @@ private:
     
     template <typename T_Src>
     void write(uint8_t epAddr, T_Src& src, Milliseconds timeout=Forever) {
-        write((uint8_t)epAddr, (void*)&src, sizeof(src), timeout);
+        write(epAddr, (void*)&src, sizeof(src), timeout);
     }
     
     void write(uint8_t epAddr, const void* buf, size_t len, Milliseconds timeout=Forever) {
-        _claimInterfaceForEndpointAddr((uint8_t)epAddr);
+        _claimInterfaceForEndpointAddr(epAddr);
         
         int xferLen = 0;
-        int ir = libusb_bulk_transfer(_handle, (uint8_t)epAddr, (uint8_t*)buf, (int)len, &xferLen,
+        int ir = libusb_bulk_transfer(_handle, epAddr, (uint8_t*)buf, (int)len, &xferLen,
             _LibUSBTimeoutFromMs(timeout));
         _CheckErr(ir, "libusb_bulk_transfer failed");
         if ((size_t)xferLen != len)
@@ -555,8 +555,8 @@ private:
     }
     
     void reset(uint8_t epAddr) {
-        _claimInterfaceForEndpointAddr((uint8_t)epAddr);
-        int ir = libusb_clear_halt(_handle, (uint8_t)epAddr);
+        _claimInterfaceForEndpointAddr(epAddr);
+        int ir = libusb_clear_halt(_handle, epAddr);
         _CheckErr(ir, "libusb_clear_halt failed");
     }
     
@@ -656,7 +656,7 @@ private:
 public:
     
     uint16_t maxPacketSize(uint8_t epAddr) const {
-        const _EndpointInfo& epInfo = _epInfo((uint8_t)epAddr);
+        const _EndpointInfo& epInfo = _epInfo(epAddr);
         return epInfo.maxPacketSize;
     }
